+++ mitra-models/src/profiles/types.rs
@@ -138,6 +138,7 @@ pub enum IdentityProofType {
     LegacyMinisignIdentityProof,
     FepC390JcsBlake2Ed25519Proof, // MitraJcsEd25519Signature2022
     FepC390JcsEip191Proof, // MitraJcsEip191Signature2022
+    FepC390JcsEddsaProof, // jcs-eddsa-2022
 }
 
 impl From<&IdentityProofType> for i16 {
@@ -147,6 +148,7 @@ impl From<&IdentityProofType> for i16 {
             IdentityProofType::LegacyMinisignIdentityProof => 2,
             IdentityProofType::FepC390JcsBlake2Ed25519Proof => 3,
             IdentityProofType::FepC390JcsEip191Proof => 4,
+            IdentityProofType::FepC390JcsEddsaProof => 5,
         }
     }
 }
@@ -160,6 +162,7 @@ impl TryFrom<i16> for IdentityProofType {
             2 => Self::LegacyMinisignIdentityProof,
             3 => Self::FepC390JcsBlake2Ed25519Proof,
             4 => Self::FepC390JcsEip191Proof,
+            5 => Self::FepC390JcsEddsaProof,
             _ => return Err(DatabaseTypeError),
         };
         Ok(proof_type)